Struct google_accesscontextmanager1_beta::AccessPolicyMethods

pub struct AccessPolicyMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on accessPolicy resources. It is not used directly, but through the AccessContextManager hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_accesscontextmanager1_beta as accesscontextmanager1_beta;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use accesscontextmanager1_beta::AccessContextManager;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AccessContextManager::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `access_levels_create(...)`, `access_levels_delete(...)`, `access_levels_get(...)`, `access_levels_list(...)`, `access_levels_patch(...)`, `create(...)`, `delete(...)`, `get(...)`, `list(...)`, `patch(...)`, `service_perimeters_create(...)`, `service_perimeters_delete(...)`, `service_perimeters_get(...)`, `service_perimeters_list(...)` and `service_perimeters_patch(...)`
// to build up your call.
let rb = hub.access_policies();

Methods

impl<'a, C, A> AccessPolicyMethods<'a, C, A>

pub fn access_levels_list(
&self,
parent: &str
) -> AccessPolicyAccessLevelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all Access Levels for an access policy.

Arguments

parent - Required. Resource name for the access policy to list Access Levels from. Format: accessPolicies/{policy_id}

pub fn access_levels_create(
&self,
request: AccessLevel,
parent: &str
) -> AccessPolicyAccessLevelCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create an Access Level. The longrunning operation from this RPC will have a successful status once the Access Level has propagated to long-lasting storage. Access Levels containing errors will result in an error response for the first error encountered.

Arguments

request - No description provided.

parent - Required. Resource name for the access policy which owns this Access Level. Format: accessPolicies/{policy_id}

pub fn service_perimeters_create(
&self,
request: ServicePerimeter,
parent: &str
) -> AccessPolicyServicePerimeterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create an Service Perimeter. The longrunning operation from this RPC will have a successful status once the Service Perimeter has propagated to long-lasting storage. Service Perimeters containing errors will result in an error response for the first error encountered.

Arguments

request - No description provided.

parent - Required. Resource name for the access policy which owns this Service Perimeter. Format: accessPolicies/{policy_id}

pub fn service_perimeters_delete(
&self,
name: &str
) -> AccessPolicyServicePerimeterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete an Service Perimeter by resource name. The longrunning operation from this RPC will have a successful status once the Service Perimeter has been removed from long-lasting storage.

Arguments

name - Required. Resource name for the Service Perimeter. Format: accessPolicies/{policy_id}/servicePerimeters/{service_perimeter_id}

pub fn access_levels_patch(
&self,
request: AccessLevel,
name: &str
) -> AccessPolicyAccessLevelPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update an Access Level. The longrunning operation from this RPC will have a successful status once the changes to the Access Level have propagated to long-lasting storage. Access Levels containing errors will result in an error response for the first error encountered.

Arguments

request - No description provided.

name - Required. Resource name for the Access Level. The short_name component must begin with a letter and only include alphanumeric and '_'. Format: accessPolicies/{policy_id}/accessLevels/{short_name}

pub fn create(&self, request: AccessPolicy) -> AccessPolicyCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create an AccessPolicy. Fails if this organization already has a AccessPolicy. The longrunning Operation will have a successful status once the AccessPolicy has propagated to long-lasting storage. Syntactic and basic semantic errors will be returned in metadata as a BadRequest proto.

Arguments

request - No description provided.

pub fn get(&self, name: &str) -> AccessPolicyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get an AccessPolicy by name.

Arguments

name - Required. Resource name for the access policy to get. Format accessPolicies/{policy_id}

pub fn list(&self) -> AccessPolicyListCall<'a, C, A>

Create a builder to help you perform the following task:

List all AccessPolicies under a container.

pub fn access_levels_get(
&self,
name: &str
) -> AccessPolicyAccessLevelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get an Access Level by resource name.

Arguments

name - Required. Resource name for the Access Level. Format: accessPolicies/{policy_id}/accessLevels/{access_level_id}

pub fn patch(
&self,
request: AccessPolicy,
name: &str
) -> AccessPolicyPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update an AccessPolicy. The longrunning Operation from this RPC will have a successful status once the changes to the AccessPolicy have propagated to long-lasting storage. Syntactic and basic semantic errors will be returned in metadata as a BadRequest proto.

Arguments

request - No description provided.

name - Output only. Resource name of the AccessPolicy. Format: accessPolicies/{policy_id}

pub fn access_levels_delete(
&self,
name: &str
) -> AccessPolicyAccessLevelDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete an Access Level by resource name. The longrunning operation from this RPC will have a successful status once the Access Level has been removed from long-lasting storage.

Arguments

name - Required. Resource name for the Access Level. Format: accessPolicies/{policy_id}/accessLevels/{access_level_id}

pub fn service_perimeters_list(
&self,
parent: &str
) -> AccessPolicyServicePerimeterListCall<'a, C, A>

Create a builder to help you perform the following task:

List all Service Perimeters for an access policy.

Arguments

parent - Required. Resource name for the access policy to list Service Perimeters from. Format: accessPolicies/{policy_id}

pub fn service_perimeters_get(
&self,
name: &str
) -> AccessPolicyServicePerimeterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get an Service Perimeter by resource name.

Arguments

name - Required. Resource name for the Service Perimeter. Format: accessPolicies/{policy_id}/servicePerimeters/{service_perimeters_id}

pub fn delete(&self, name: &str) -> AccessPolicyDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete an AccessPolicy by resource name. The longrunning Operation will have a successful status once the AccessPolicy has been removed from long-lasting storage.

Arguments

name - Required. Resource name for the access policy to delete. Format accessPolicies/{policy_id}

pub fn service_perimeters_patch(
&self,
request: ServicePerimeter,
name: &str
) -> AccessPolicyServicePerimeterPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update an Service Perimeter. The longrunning operation from this RPC will have a successful status once the changes to the Service Perimeter have propagated to long-lasting storage. Service Perimeter containing errors will result in an error response for the first error encountered.

Arguments

request - No description provided.

name - Required. Resource name for the ServicePerimeter. The short_name component must begin with a letter and only include alphanumeric and '_'. Format: accessPolicies/{policy_id}/servicePerimeters/{short_name}